--- src/components/constants.js
@@ -1,7 +1,0 @@
-export const POINTS_MARKER_COLOR = '#098CEA';
-export const RED_MARKER_COLOR = '#FF0000';
-
-export const PAPERCOLORS = [
-  0x1aa24d, 0xf2482e, 0x343394, 0xfed53d, 0x000000, 0xff9fd5, 0x9ddef4,
-  0xff8158,
-];
