+++ src/constants/index.js
@@ -1,0 +1,8 @@
+const POINTS_MARKER_COLOR = '#098CEA';
+const RED_MARKER_COLOR = '#FF0000';
+const PAPERCOLORS = [
+  0x1aa24d, 0xf2482e, 0x343394, 0xfed53d, 0x000000, 0xff9fd5, 0x9ddef4,
+  0xff8158,
+];
+
+export { POINTS_MARKER_COLOR, RED_MARKER_COLOR, PAPERCOLORS };
